Moved test scaffolding to top level
